refactor(web): collapse padding axes and split cva variants into own files
- px-N py-N -> p-N where axes match (12)
- move button/badge/tabs/callout cva variants to *-variants.ts so the
  component modules only export components (fixes only-export-components)
