+++ apps/web/src/components/journal/journal-card.tsx
@@ -72,7 +72,7 @@ export function JournalCard({
                     <button
                       type="button"
                       onClick={() => onEdit(entry)}
-                      className="flex w-full items-center gap-2 rounded-md px-2 py-2 text-sm hover:bg-accent focus-visible:bg-accent focus-visible:outline-none"
+                      className="flex w-full items-center gap-2 rounded-md p-2 text-sm hover:bg-accent focus-visible:bg-accent focus-visible:outline-none"
                     >
                       <Pencil className="size-4 text-muted-foreground" />
                       {t("child.edit")}
@@ -82,7 +82,7 @@ export function JournalCard({
                     <button
                       type="button"
                       onClick={() => onDelete(entry)}
-                      className="flex w-full items-center gap-2 rounded-md px-2 py-2 text-sm text-destructive hover:bg-destructive/10 focus-visible:bg-destructive/10 focus-visible:outline-none"
+                      className="flex w-full items-center gap-2 rounded-md p-2 text-sm text-destructive hover:bg-destructive/10 focus-visible:bg-destructive/10 focus-visible:outline-none"
                     >
                       <Trash2 className="size-4" />
                       {t("child.delete")}
