+++ apps/web/src/components/ui/callout.tsx
@@ -1,26 +1,9 @@
 import * as React from "react";
-import { cva, type VariantProps } from "class-variance-authority";
+import { type VariantProps } from "class-variance-authority";
 import { Lightbulb, AlertTriangle, CheckCircle2, Info, type LucideIcon } from "lucide-react";
 
 import { cn } from "@/lib/utils";
-
-const calloutVariants = cva(
-  "flex gap-3 rounded-lg border px-4 py-3 text-sm leading-relaxed",
-  {
-    variants: {
-      variant: {
-        info: "border-info-border bg-info-surface text-info-foreground",
-        tip: "border-info-border bg-info-surface text-info-foreground",
-        warning: "border-warning-border bg-warning-surface text-warning-foreground",
-        success: "border-success-border bg-success-surface text-success-foreground",
-        danger: "border-danger-border bg-danger-surface text-danger-foreground",
-      },
-    },
-    defaultVariants: {
-      variant: "info",
-    },
-  },
-);
+import { calloutVariants } from "./callout-variants";
 
 const defaultIcons: Record<NonNullable<VariantProps<typeof calloutVariants>["variant"]>, LucideIcon> = {
   info: Info,
@@ -47,4 +30,4 @@ function Callout({ className, variant, icon, children, ...props }: CalloutProps)
   );
 }
 
-export { Callout, calloutVariants };
+export { Callout };
