+++ apps/web/src/routes/ressources/plan-de-crise.tsx
@@ -33,7 +33,7 @@ function PlanDeCrisePage() {
         <div className="plan-de-crise-page min-h-dvh bg-background">
             {/* Screen-only controls */}
             <div className="plan-de-crise-controls border-b border-border/60 bg-background/95 backdrop-blur-sm">
-                <div className="mx-auto flex max-w-3xl flex-wrap items-center justify-between gap-3 px-4 py-4">
+                <div className="mx-auto flex max-w-3xl flex-wrap items-center justify-between gap-3 p-4">
                     <Link
                         to="/ressources/$slug"
                         params={{ slug: "crise-tdah-enfant-guide-complet" }}
